refactor: 고객 banking 상태 hook 분리
